Separate the text editor from file itself
